Remove "go1.10" build tag (#1016)
Some IDEs and editors refuse to acknowledge the "go1.10" build tag when
autocompleting & compiling. Removing said tag increases usibility of the
library for those stuck with these editors.
